CHANGEMENT DES POINT H ET CPP
