+++ src/main.cpp
@@ -15,7 +15,7 @@
 
 #include <Arduino.h>
 #include <iostream>
-#include "Etudiant.h"
+#include "EtudiantTGE.h"
 
 //********************************************************************************
 // Déclaration des fonctions
